Change name login test
