+++ qa-automation/src/test/java/com/digitalbank/qaautomation/tests/LoginRememberMeTest.java
@@ -1,6 +1,8 @@
 package com.digitalbank.qaautomation.tests;
 
-import com.digitalbank.qaautomation.dataProviders.ShouldNotLoginData;
+
+import com.digitalbank.qaautomation.dataProviders.RememberMeLoginData;
+import com.digitalbank.qaautomation.pages.HeaderPage;
 import io.qameta.allure.Description;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
@@ -11,7 +13,8 @@ import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 import com.digitalbank.qaautomation.pages.LoginPage;
 
-public class TryLoginWithWrongPassword {
+public class LoginRememberMeTest {
+
 
     private WebDriver driver;
 
@@ -28,21 +31,28 @@ public class TryLoginWithWrongPassword {
         driver.quit();
     }
 
-    @Description("This test should try to log in to the website with valid username and wrong password")
-    @Test(testName = "Try login with valid username and wrong password ",dataProvider = "credentials", dataProviderClass = ShouldNotLoginData.class)
-    public void shouldNotLoginWhenWrongPasswordIsProvided(String user, String pass) throws Exception {
+    @Description("This test will verify that after logout the 'Username' credential is remembered if the 'Remember Me' box was checked on login. ")
+    @Test(testName = "Try login with valid username and wrong password ",dataProvider = "credentials", dataProviderClass = RememberMeLoginData.class)
+    public void shouldRembemberLastLogedUsernameWhenRememberMeCheckboxIsCheck(String user, String pass) throws Exception {
 
         driver.get("http://digitalbank.upcamp.io/bank/login");
 
         LoginPage loginPage = new LoginPage(driver);
         loginPage.logIn(user, pass);
+        HeaderPage headerPage = new HeaderPage(driver);
 
         String currentUrl = driver.getCurrentUrl();
-        String expectedUrl = "http://digitalbank.upcamp.io/bank/login?error";
+        String expectedUrl = "http://digitalbank.upcamp.io/bank/home";
         Assert.assertEquals(currentUrl, expectedUrl);
 
-        String cantLoginAlert = loginPage.getCantLoginAlert();
-        String expectedCantLoginAlert = ("Error Invalid credentials or access not granted due to user account status or an existing user session.\n" + "×");
-        Assert.assertEquals(cantLoginAlert, expectedCantLoginAlert);
+        headerPage.logOut();
+
+        String currentLogoutUrl = driver.getCurrentUrl();
+        String expectedLogoutUrl = "http://digitalbank.upcamp.io/bank/login";
+        Assert.assertEquals(currentLogoutUrl, expectedLogoutUrl);
+        String rememberedUsername = loginPage.getRememberedUsername();
+        Assert.assertEquals(rememberedUsername, user);
+        
     }
 }
+
